Add support for Tsuru Java platform
This commit adds two files to create a build hook for tsuru java platform.

These changes allow the Tsuru Java platform to build the java project inside the platform. Deploying
the built jar (built using maven) can also be used as a deployment process. This is documented
in the Ops manual.