Tsuru healthcheck config for JAR
I added this to the root of the repo in 2fb4d0b but that isn't used when
deploying the application from a JAR with `tsuru app-deploy`. We can't
symlink the file like we did in c4b77ae because we don't want the `hooks`
section, so we have to create a separate copy.