Modify pom.xml to generate a uber jar for CF
In order to run this application in Cloudfoundry, we need to follow
the [instructions for a single main
app](https://github.com/cloudfoundry/java-buildpack/blob/master/docs/container-java_main.md) described in the
[cloudfoundry java
buildpack](https://github.com/cloudfoundry/java-buildpack).

We follow the pattern of
[uber-jar](http://stackoverflow.com/questions/11947037/what-is-an-uber-jar)
creating a big jar which includes the code and all the required jars,
and we specify the classpath and main class in the Manifest of the main
jar.

In this commit we [implement this
solution](http://stackoverflow.com/a/25116745/395686), which consists
in:
 1. Copy all dependency jars to `target/classes/lib` when building using
   the `maven-dependency-plugin`
 2. Modify the manifest using `maven-jar-plugin` to include the
 classpath of the dependencies in `lib/` and include the main class

Other approaches can be use `maven-assembly-plugin` or
`onejar-maven-plugin`.